requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
ecdsa==0.19.0
exceptiongroup==1.2.1
fastapi==0.110.2
h11==0.14.0
idna==3.7
jwt==1.3.1
MouseInfo==0.1.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pyasn1==0.6.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
python-jose==3.3.0
python-multipart==0.0.9
pytweening==1.2.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
